Fixing bug running tests, changing db names, can now pass db on command line...
The following database names can be passed:
circle-cloud, circle-local, gae-cloud, dev-cloud, dev-local

The names should be self-explanatory. gae-cloud is the default in the absense of a setting.

The gaetestrunner command line is now:

python gaetestrunner.py path-to-GAE-SDK [+|-test-filter] [db-env-name]'

Filters must be preceded by a "+" (positive filter, "run only this test")'
or a "-", negative filter "run all but this test"'
currently, only one filter is supported'